{long,int} -> int32_t

+++ include/protocols/rwhod.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: rwhod.h,v 1.3 1994/10/26 00:56:51 cgd Exp $	*/
+/*	$NetBSD: rwhod.h,v 1.4 1996/09/23 05:01:08 mycroft Exp $	*/
 
 /*
  * Copyright (c) 1983 The Regents of the University of California.
@@ -44,21 +44,21 @@
 struct	outmp {
 	char	out_line[8];		/* tty name */
 	char	out_name[8];		/* user id */
-	long	out_time;		/* time on */
+	int32_t	out_time;		/* time on */
 };
 
 struct	whod {
 	char	wd_vers;		/* protocol version # */
 	char	wd_type;		/* packet type, see below */
 	char	wd_pad[2];
-	int	wd_sendtime;		/* time stamp by sender */
-	int	wd_recvtime;		/* time stamp applied by receiver */
+	int32_t	wd_sendtime;		/* time stamp by sender */
+	int32_t	wd_recvtime;		/* time stamp applied by receiver */
 	char	wd_hostname[32];	/* hosts's name */
-	int	wd_loadav[3];		/* load average as in uptime */
-	int	wd_boottime;		/* time system booted */
+	int32_t	wd_loadav[3];		/* load average as in uptime */
+	int32_t	wd_boottime;		/* time system booted */
 	struct	whoent {
 		struct	outmp we_utmp;	/* active tty info */
-		int	we_idle;	/* tty idle time */
+		int32_t	we_idle;	/* tty idle time */
 	} wd_we[1024 / sizeof (struct whoent)];
 };
